+++ include/cc_sakura.h
@@ -15,6 +15,7 @@ typedef enum{
 	TK_ELSE,
 	TK_WHILE,
 	TK_FOR,
+	TK_SIZEOF,
 	TK_BLOCK,
 	TK_RETURN,
 	TK_EOF,
@@ -101,26 +102,28 @@ struct LVar{
 	LVar *next;
 };
 
+// tokenizer
+int len_val(char *str);
+bool issymbol(char *str, bool *flag);
+bool isblock(char *str);
+bool at_eof();
+Token *tokenize(char *p);
+Token *new_token(TokenKind kind,Token *cur,char *str);
+
 // parse.c
 int lvar_count;
+char *user_input;
 Token *token;
 Func *func_list[100];
 LVar *locals;
 
-int len_val(char *str);
-char *user_input;
 void error(char *loc,char *fmt, ...);
-bool issymbol(char *str, bool *flag);
-bool isblock(char *str);
-bool at_eof();
 bool consume(char *op);
 bool consume_ret();
 bool consume_reserved_word();
 void expect(char *op);
 int expect_number();
 
-Token *new_token(TokenKind kind,Token *cur,char *str);
-Token *tokenize(char *p);
 Token *consume_ident();
 
 Node *new_node(NodeKind kind,Node *lhs,Node *rhs);
